Decouple AnalysisConfigGenerator from coverage post processing
